+++ src/args.rs
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 use std::fmt::Debug;
+use std::path::PathBuf;
 
 use clap::Parser;
 
@@ -10,6 +11,12 @@ use clap::Parser;
 #[derive(Debug, Parser)]
 #[clap(version = env!("VERSION"))]
 pub struct Args {
+  /// The configuration directory to use.
+  ///
+  /// The directory typically contains a `notnow.json` configuration
+  /// alongside all tasks in the `tasks/` sub-directory.
+  #[clap(short, long)]
+  pub config_dir: Option<PathBuf>,
   /// Force reclamation of stale lock files in case a previous program
   /// instance terminated improperly.
   #[clap(short, long)]